New Classes Model & Transform
Appearance aspects of scene objects are now in a model class and all transform information in contained in the transform class. The scene object now takes in a pointer to a model & a transform. This composite structure is more flexible than the previous design. Transform has a virtual update method which performing all transformations; it can be overridden to perform different transformations than the default, The SGnode class has also been amalgamated with scene object. Scene object has a statically allocated 5 children per scene object node.